requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
